Version 1.0
Takes NetCDF data, looks at metadata, dimensions, and variables. Converts time values into a Python datetime format and can select a specific date to plot precipitation for.

No Cartopy functionality as that section of the tutorial seems to be deprecated (see Troubleshooting branch where the original tutorial data was used and did not work with the code).